+++ misc/generate_keyword_hash_fn.sh
@@ -120,6 +120,7 @@ var
 varargs
 virtual
 while
+winapi
 with
 write
 writeonly